feat: seperate purchased clips component
